# openapi.json
{"openapi":"3.1.0","info":{"title":"AI电商评论分析服务","description":"对电商用户评论进行AI驱动的结构化分析，支持多用户配额管理","version":"1.1.0"},"paths":{"/admin/generate-key":{"post":{"tags":["管理"],"summary":"Generate Key","description":"生成新的 API Key。\n需要超级管理员密钥授权: `Authorization: Bearer <ADMIN_API_KEY>`","operationId":"generate_key_admin_generate_key_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateKeyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/list-keys":{"get":{"tags":["管理"],"summary":"List All Keys","description":"列出所有 API Key（需要管理员授权）","operationId":"list_all_keys_admin_list_keys_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/reset-calls":{"post":{"tags":["管理"],"summary":"Reset Calls","description":"重置月度调用计数。\n不传 api_key 则重置全部用户。","operationId":"reset_calls_admin_reset_calls_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Health Check","description":"根路径 —— 返回健康检查信息","operationId":"health_check__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analyze-reviews":{"post":{"summary":"Analyze Reviews","description":"分析电商用户评论。\n\n鉴权:  Bearer Token（通过管理员接口 /admin/generate-key 生成）\n限流:  根据套餐月度配额，超限返回 429\n\n返回 AI 驱动的结构化分析结果。","operationId":"analyze_reviews_api_v1_analyze_reviews_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnalyzeRequest":{"properties":{"product_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Name","description":"商品名称（可选）","example":"无线蓝牙耳机 Pro Max"},"reviews":{"items":{"type":"string"},"type":"array","minItems":1,"title":"Reviews","description":"用户评论列表，每条是一条用户评论","example":["音质很好，续航也不错","戴久了耳朵疼，降噪一般"]},"max_reviews":{"type":"integer","maximum":300.0,"minimum":1.0,"title":"Max Reviews","description":"最多分析的评论数量，默认100，最大300","default":100,"example":50}},"type":"object","required":["reviews"],"title":"AnalyzeRequest","description":"评论分析请求体"},"GenerateKeyRequest":{"properties":{"user_id":{"type":"string","minLength":1,"title":"User Id","description":"用户ID","example":"user_001"},"api_key":{"type":"string","minLength":8,"title":"Api Key","description":"自定义 API Key 字符串","example":"sk-my-custom-key-123"},"plan":{"type":"string","title":"Plan","description":"套餐类型","default":"free","example":"basic"},"call_limit_per_month":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Call Limit Per Month","description":"自定义月调用上限（不传则按套餐默认值）","example":1000}},"type":"object","required":["user_id","api_key"],"title":"GenerateKeyRequest","description":"生成新 API Key 的请求体"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ResetRequest":{"properties":{"api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Key","description":"要重置的 API Key，不传则重置全部"}},"type":"object","title":"ResetRequest","description":"重置月度计数的请求体"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}